requirements.txt
allure_python_commons==2.9.41
clickhouse_sqlalchemy==0.2.1
Flask==2.2.2
jsonpath==0.82
pandas==1.3.4
paramiko==2.10.3
pymongo==3.11.4
PyMySQL==1.0.2
PyQt5==5.15.9
pytest==6.2.4
PyYAML==6.0
redis==3.5.3
requests==2.25.1
SQLAlchemy==1.4.40
urllib3==1.26.4
faker==18.4.0
